Split match_predicate file
